refactor(cli): split tsserver structure refactor handlers
